refactor(titlebar): move plus button to WMUX right side
Move the add-workspace button from StatusBar back to Titlebar left segment,
positioning it immediately right of the WMUX wordmark. Remove it from StatusBar.
